backend/requirements.txt
# Core compatibility requirements
# Generated from pyproject.toml by backend/scripts/export_requirements.py
# Do not edit manually.

fastapi>=0.115,<1.0
sqlalchemy>=2.0,<3.0
pydantic>=2.8,<3.0
pandas>=2.2,<3.0
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
akshare>=1.18,<2.0
pybind11>=3.0,<4.0
python-dotenv>=1.0,<2.0
alembic>=1.16,<2.0
psycopg[binary]>=3.2,<4.0
uvicorn>=0.34,<1.0
httpx>=0.28,<1.0
